// extensions/discounts-allocator/generated/api.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  /**
   * Represents an [ISO 8601](https://en.wikipedia.org/wiki/ISO_8601)-encoded date string.
   * For example, September 7, 2019 is represented as `"2019-07-16"`.
   */
  Date: { input: any; output: any; }
  /**
   * Represents an [ISO 8601](https://en.wikipedia.org/wiki/ISO_8601)-encoded date and time string.
   * For example, 3:50 pm on September 7, 2019 in the time zone of UTC (Coordinated Universal Time) is
   * represented as `"2019-09-07T15:50:00Z`".
   */
  DateTime: { input: any; output: any; }
  /**
   * A subset of the [ISO 8601](https://en.wikipedia.org/wiki/ISO_8601) format that
   * includes the date and time but not the timezone which is determined from context.
   *
   * For example, "2018-01-01T00:00:00".
   */
  DateTimeWithoutTimezone: { input: any; output: any; }
  /**
   * A signed decimal number, which supports arbitrary precision and is serialized as a string.
   *
   * Example values: `"29.99"`, `"29.999"`.
   */
  Decimal: { input: any; output: any; }
  /**
   * A function-scoped handle to a refer a resource.
   * The Handle type appears in a JSON response as a String, but it is not intended to be human-readable.
   * Example value: `"10079785100"`
   */
  Handle: { input: any; output: any; }
  /**
   * A [JSON](https://www.json.org/json-en.html) object.
   *
   * Example value:
   * `{
   *   "product": {
   *     "id": "gid://shopify/Product/1346443542550",
   *     "title": "White T-shirt",
   *     "options": [{
   *       "name": "Size",
   *       "values": ["M", "L"]
   *     }]
   *   }
   * }`
   */
  JSON: { input: any; output: any; }
  /**
   * A subset of the [ISO 8601](https://en.wikipedia.org/wiki/ISO_8601) format that
   * includes the time but not the date or timezone which is determined from context.
   * For example, "05:43:21".
   */
  TimeWithoutTimezone: { input: any; output: any; }
  /** A void type that can be used to return a null value from a mutation. */
  Void: { input: any; output: any; }
};
// ...
export type Cart = {
  __typename?: 'Cart';
  /** The attributes associated with the cart. Attributes are represented as key-value pairs. */
  attribute?: Maybe<Attribute>;
  /** Information about the buyer that is interacting with the cart. */
  buyerIdentity?: Maybe<BuyerIdentity>;
  /** The costs that the buyer will pay at checkout. */
  cost: CartCost;
  /** A list of lines containing information about the items that can be delivered. */
  deliverableLines: Array<DeliverableCartLine>;
  /** The delivery groups available for the cart based on the buyer's shipping address. */
  deliveryGroups: Array<CartDeliveryGroup>;
  /** A list of lines containing information about the items the customer intends to purchase. */
  lines: Array<CartLine>;
  /** The localized fields available for the cart. */
  localizedFields: Array<LocalizedField>;
};
// ...
export type CartLine = {
  __typename?: 'CartLine';
  /**
   * Retrieve a cart line attribute by key.
   *
   * Cart line attributes are also known as line item properties in Liquid.
   */
  attribute?: Maybe<Attribute>;
  /** The cost of the merchandise line that the buyer will pay at checkout. */
  cost: CartLineCost;
  /** The ID of the cart line. */
  id: Scalars['ID']['output'];
  /** The merchandise that the buyer intends to purchase. */
  merchandise: Merchandise;
  /** The quantity of the merchandise that the customer intends to purchase. */
  quantity: Scalars['Int']['output'];
  /**
   * The selling plan associated with the cart line and the effect that each
   * selling plan has on variants when they're purchased.
   */
  sellingPlanAllocation?: Maybe<SellingPlanAllocation>;
};


/** Represents information about the merchandise in the cart. */
export type CartLineAttributeArgs = {
  key?: InputMaybe<Scalars['String']['input']>;
};

/** The cost of the merchandise line that the buyer will pay at checkout. */
export type CartLineCost = {
  __typename?: 'CartLineCost';
  /** The amount of the merchandise line. */
  amountPerQuantity: MoneyV2;
  /**
   * The compare at amount of the merchandise line. This value varies depending on
   * the buyer's identity, and is null when the value is hidden to buyers.
   */
  compareAtAmountPerQuantity?: Maybe<MoneyV2>;
  /** The cost of the merchandise line before line-level discounts. */
  subtotalAmount: MoneyV2;
  /** The total cost of the merchandise line. */
  totalAmount: MoneyV2;
};

export type CartLineTarget = {
  __typename?: 'CartLineTarget';
  /** The ID of the line to be discounted. */
  cartLineId: Scalars['ID']['output'];
  /** The quantity of the items on this line to be discounted. */
  quantity: Scalars['Int']['output'];
};
// ...
export type CustomProduct = {
  __typename?: 'CustomProduct';
  /** Whether the merchandise is a gift card. */
  isGiftCard: Scalars['Boolean']['output'];
  /** Whether the merchandise requires shipping. */
  requiresShipping: Scalars['Boolean']['output'];
  /** The localized title of the product in the customer’s locale. */
  title: Scalars['String']['output'];
  /** The weight of the product variant in the unit system specified with `weight_unit`. */
  weight?: Maybe<Scalars['Float']['output']>;
  /** Unit of measurement for weight. */
  weightUnit: WeightUnit;
};
// ...
export type Discount = HasMetafields & {
  __typename?: 'Discount';
  /** The discount code, for code-based discounts. */
  code?: Maybe<Scalars['String']['output']>;
  /** The strategy to use when considering discounts proposals. */
  discountApplicationStrategy: DiscountApplicationStrategy;
  /** The class of the discount. */
  discountClass: DiscountClass;
  /** Proposals to evaluate as part of this discount. */
  discountProposals: Array<DiscountProposal>;
  /** The ID of the discount. */
  id: Scalars['ID']['output'];
  /** Returns a metafield by namespace and key that belongs to the resource. */
  metafield?: Maybe<Metafield>;
  /** The discount title. */
  title: Scalars['String']['output'];
};


export type DiscountMetafieldArgs = {
  key: Scalars['String']['input'];
  namespace?: InputMaybe<Scalars['String']['input']>;
};

/** The strategy that's applied to the list of discounts. */
export enum DiscountApplicationStrategy {
  /** Apply all discounts with conditions that are satisfied. This does not override discount combination or stacking rules. */
  All = 'ALL',
  /** Only apply the first discount with conditions that are satisfied. */
  First = 'FIRST',
  /** Only apply the discount that offers the maximum reduction. */
  Maximum = 'MAXIMUM'
}

/** List of the classes of discounts. */
export enum DiscountClass {
  /** A discount that applies to the order. */
  Order = 'ORDER',
  /** A discount that applies to products. */
  Product = 'PRODUCT'
}

/** The input object for the Function. */
export type DiscountProposal = {
  __typename?: 'DiscountProposal';
  /** The handle of the discount proposal. */
  handle: Scalars['Handle']['output'];
  /** The message for this proposal. */
  message?: Maybe<Scalars['String']['output']>;
  /** The line items to which the discount proposal applies. */
  targets: Array<CartLineTarget>;
  /** The value of the discount. */
  value: Value;
};

export type DisplayableError = {
  /** The ID of the discount that caused this violation. */
  discountId: Scalars['ID']['input'];
  /** The error message in the user's locale. */
  reason: Scalars['String']['input'];
};

/** A fixed amount value. */
export type FixedAmount = {
  __typename?: 'FixedAmount';
  /** The fixed amount value of the discount, in the currency of the cart. */
  amount: Scalars['Decimal']['output'];
  /**
   * Allocation method of the discount, will be allocated on each line if it's
   * true, else will be allocated across target lines.
   */
  appliesToEachItem: Scalars['Boolean']['output'];
};

/** The run target result. */
export type FunctionRunResult = {
  /** The list of displayable errors. */
  displayableErrors?: InputMaybe<Array<DisplayableError>>;
  /** The list of applied line discounts. */
  lineDiscounts?: InputMaybe<Array<LineDiscount>>;
};
// ...
export type HasMetafields = {
  /** Returns a metafield by namespace and key that belongs to the resource. */
  metafield?: Maybe<Metafield>;
};
// ...
export type Input = {
  __typename?: 'Input';
  /** The cart. */
  cart: Cart;
  /** The list of discounts to be allocated. */
  discounts?: Maybe<Array<Discount>>;
  /** The localization of the Function execution context. */
  localization: Localization;
  /** The conversion rate between the shop's currency and the currency of the cart. */
  presentmentCurrencyRate: Scalars['Decimal']['output'];
  /** Information about the shop. */
  shop: Shop;
};
// ...
export type LineDiscount = {
  /** The output allocations for this line discount. */
  allocations: Array<OutputAllocations>;
  /** The ID of the cart line associated to this discount. */
  cartLineId: Scalars['ID']['input'];
  /** The quantity to which this line discount will apply. */
  quantity: Scalars['Int']['input'];
};
// ...
export type Localization = {
  __typename?: 'Localization';
  /** The country of the active localized experience. */
  country: Country;
  /** The language of the active localized experience. */
  language: Language;
  /** The market of the active localized experience. */
  market: Market;
};
// ...
export type Merchandise = CustomProduct | ProductVariant;
// ...
export type Metafield = {
  __typename?: 'Metafield';
  /** The data stored in the metafield in JSON format. */
  jsonValue: Scalars['JSON']['output'];
  /**
   * The type of data that the metafield stores in the `value` field.
   * Refer to the list of [supported types](https://shopify.dev/apps/metafields/types).
   */
  type: Scalars['String']['output'];
  /** The data stored in the metafield. Always stored as a string, regardless of the metafield's type. */
  value: Scalars['String']['output'];
};

/** A monetary value with currency. */
export type MoneyV2 = {
  __typename?: 'MoneyV2';
  /** Decimal money amount. */
  amount: Scalars['Decimal']['output'];
  /** Currency of the money. */
  currencyCode: CurrencyCode;
};
// ...
export type Percentage = {
  __typename?: 'Percentage';
  /** The percentage value. */
  value: Scalars['Decimal']['output'];
};
// ...
export type ProductVariant = HasMetafields & {
  __typename?: 'ProductVariant';
  /** A globally-unique identifier. */
  id: Scalars['ID']['output'];
  /** Returns a metafield by namespace and key that belongs to the resource. */
  metafield?: Maybe<Metafield>;
  /** The product that this variant belongs to. */
  product: Product;
  /** Whether the merchandise requires shipping. */
  requiresShipping: Scalars['Boolean']['output'];
  /** An identifier for the product variant in the shop. Required in order to connect to a fulfillment service. */
  sku?: Maybe<Scalars['String']['output']>;
  /** The localized title of the product variant in the customer’s locale. */
  title?: Maybe<Scalars['String']['output']>;
  /** The weight of the product variant in the unit system specified with `weight_unit`. */
  weight?: Maybe<Scalars['Float']['output']>;
  /** Unit of measurement for weight. */
  weightUnit: WeightUnit;
};
// ...
export type Shop = HasMetafields & {
  __typename?: 'Shop';
  /** Information about the current time relative to the shop's timezone setting. */
  localTime: LocalTime;
  /** Returns a metafield by namespace and key that belongs to the resource. */
  metafield?: Maybe<Metafield>;
};


/** Information about the shop. */
export type ShopMetafieldArgs = {
  key: Scalars['String']['input'];
  namespace?: InputMaybe<Scalars['String']['input']>;
};

/** The value of the discount. */
export type Value = FixedAmount | Percentage;

/** Units of measurement for weight. */
export enum WeightUnit {
  /** Metric system unit of mass. */
  Grams = 'GRAMS',
  /** 1 kilogram equals 1000 grams. */
  Kilograms = 'KILOGRAMS',
  /** Imperial system unit of mass. */
  Ounces = 'OUNCES',
  /** 1 pound equals 16 ounces. */
  Pounds = 'POUNDS'
}

export type RunInputVariables = Exact<{ [key: string]: never; }>;


export type RunInput = { __typename?: 'Input', discounts?: Array<{ __typename?: 'Discount', id: string, title: string, code?: string | null, discountApplicationStrategy: DiscountApplicationStrategy, discountProposals: Array<{ __typename?: 'DiscountProposal', handle: any, message?: string | null, value: { __typename: 'FixedAmount', amount: any, appliesToEachItem: boolean } | { __typename: 'Percentage', value: any }, targets: Array<{ __typename?: 'CartLineTarget', cartLineId: string, quantity: number }> }> }> | null, cart: { __typename?: 'Cart', lines: Array<{ __typename?: 'CartLine', id: string, quantity: number, cost: { __typename?: 'CartLineCost', amountPerQuantity: { __typename?: 'MoneyV2', amount: any } }, merchandise: { __typename: 'CustomProduct' } | { __typename: 'ProductVariant', id: string, metafield?: { __typename?: 'Metafield', value: string } | null } }> } };